Move to declarative syntax for Storybook
