+++ package.json
@@ -29,15 +29,15 @@
   },
   "devDependencies": {
     "@kazupon/vue-i18n-loader": "^0.3.0",
-    "@storybook/addon-a11y": ">=5.3.0",
-    "@storybook/addon-actions": ">=5.3.0",
-    "@storybook/addon-backgrounds": ">=5.3.0",
-    "@storybook/addon-docs": ">=5.3.0",
-    "@storybook/addon-knobs": ">=5.3.0",
+    "@storybook/addon-a11y": ">=5.3.8",
+    "@storybook/addon-actions": ">=5.3.8",
+    "@storybook/addon-backgrounds": ">=5.3.8",
+    "@storybook/addon-docs": ">=5.3.8",
+    "@storybook/addon-knobs": ">=5.3.8",
     "@storybook/addon-links": ">=5.3.0",
-    "@storybook/addon-notes": ">=5.3.0",
-    "@storybook/addon-viewport": ">=5.3.0",
-    "@storybook/vue": ">=5.3.0",
+    "@storybook/addon-notes": ">=5.3.8",
+    "@storybook/addon-viewport": ">=5.3.8",
+    "@storybook/vue": ">=5.3.8",
     "@vue/cli-plugin-babel": "^4.1.2",
     "@vue/cli-plugin-eslint": "^4.1.2",
     "@vue/cli-plugin-unit-jest": "^4.1.2",
@@ -57,7 +57,7 @@
     "stylus": "^0.54.7",
     "stylus-loader": "^3.0.2",
     "vue-cli-plugin-i18n": "^0.6.0",
-    "vue-cli-plugin-storybook": "^1.0.0",
+    "vue-cli-plugin-storybook": "^1.2.0",
     "vue-cli-plugin-styleguidist": "^4.4.2",
     "vue-template-compiler": "^2.6.10"
   },
