--- storybook/addons.js
@@ -1,10 +1,0 @@
-// Addons pane
-import '@storybook/addon-knobs/register'
-import '@storybook/addon-actions/register'
-import '@storybook/addon-a11y/register'
-
-// Demo pane
-import '@storybook/addon-backgrounds/register'
-import '@storybook/addon-viewport/register'
-
-import '@storybook/addon-links/register'
