+++ storybook/manager.js
@@ -1,0 +1,7 @@
+import { addons } from '@storybook/addons'
+
+import theme from './theme'
+
+addons.setConfig({
+  theme
+})
