--- storybook/presets.js
@@ -1,3 +1,0 @@
-module.exports = [
-//   '@storybook/addon-docs/vue/preset'
-]
